+++ core/CliMulti/Lock.php
@@ -6,7 +6,10 @@
  * @link http://piwik.org
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
-namespace Piwik;
+namespace Piwik\CliMulti;
+
+use Piwik\Filesystem;
+use Piwik\SettingsServer;
 
 class Lock
 {
