+++ tests/PHPUnit/Core/CliMulti/LockTest.php
@@ -5,7 +5,7 @@
  * @link http://piwik.org
  * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
  */
-use Piwik\Lock;
+use Piwik\CliMulti\Lock;
 
 /**
  * Class LockTest
